Client - fix payable ETH value when calling the function in Solidity
Name of the variable containing the correct WEI amount to send was incorrectly calculated based on the function.
This might not be true if we link the value from another's function input in the diagram, so we need to extract it from the inputset received by the function.
There can only be one input with name containing `wei_value`. We treat this as a restricted keyword.